Move functions to once directory
